Add Email Validator test project and initial files
This commit introduces the Email Validator test project to validate our email addresses. We have added a new .NET 6.0 project called email-validator-tests with several package references for testing. Also added is the skeleton for EmailValidatorTest class with a single test case 'IsValidEmailTest', which will be fleshed out in future commits. An instruction 'in-class.md' is added to guide others about the task. Lastly, the solution file is updated to include the new project.