Remove hardcoded greeting implementations in tests
We have removed the hardcoded greeting implementations in our tests. The purpose is to make the tests simpler and focus more on the logic of the test case rather than setting up the test data. Next step will be to implement additional helper methods to tackle the string formatting within the tests.